backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
SQLAlchemy==2.0.6
Werkzeug==2.2.3
pytest==7.2.2
langchain==0.0.236
openai==0.28.1